StockPricesReport/requirements.txt
apache-airflow
apache-airflow-providers-http
apache-airflow-providers-postgres
requests
python-dotenv
pandas
seaborn
matplotlib
python-docx
pyarrow
scikit-learn